+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] 2019-23-07
+
+### Fixed
+- Fix feature viz property value
+- Fix `toString` method in animation expression
+- Fix `carto.isBrowserSupported` adding a required check on MAX_VERTEX_TEXTURE_IMAGE_UNITS
+
+### Added
+- Support dates in binary expressions
+
 ## [1.4.0] 2019-15-07
 
 ### Fixed